Admin signout authencation api integration

+++ admin-dashboard/src/store/actions/auth.actions.js
@@ -63,9 +63,25 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
+
         dispatch({
             type: authConstants.LOGOUT_REQUEST,
-        });
+        })
+
+        const res = await axios.post('/admin/signout');
+        if(res.status === 200){
+            localStorage.clear();
+            dispatch({
+                type: authConstants.LOGOUT_SUCCESS,
+            });
+        }
+        else{
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: {
+                    error: res.data.error,
+                }
+            });
+        }
     }
 }
